requirements.txt
numpy==2.1.1
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
PyOpenGL==3.1.9
Pillow==11.1.0
trimesh==4.6.6
scikit-learn==1.6.1
pyperclip==1.9.0